requirements.txt
# Use the base version with fewer dependencies
langflow-base==1.0.0a58
uvicorn[standard]==0.23.2